Full_project/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "openai-agents>=0.3.0",
    "pymongo>=4.15.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
